no funciona

+++ Kernel/memory/memory.h
@@ -2,7 +2,7 @@
 #include<inttypes.h>
 
 #define PAGE 0x1000
-void mem_setup(unsigned himem_size);
+void mem_setup(uint64_t himem_size);
 int block_size(int level);
 int level_size(int level);
 void *
